fix(globals): Use shallow copy to update the globals.params / $state.params object
Previously, the `globals.params` object was updated using a deep copy.
This causes problems when object-type parameter values aren't cloneable.

Now, the parameter values are copied to the `globals.params` object using a shallow copy, i.e., `Object.assign`.
This is consistent with parameter handling in the `PathNode` code also.

If this commit is causing problems, it is likely that app code is mutating
the existing object parameters and then calling `state.go` with the mutated
parameters.  Recommend making a clone of the params before calling `state.go`.

Closes https://github.com/ui-router/core/issues/74