feat: Add support for react-redux 6 and onwards (#10)
BREAKING CHANGE:  `@uirouter/redux` now only supports `react-redux` 6.x onwards, with the new React Context api.

Internally the library uses `ReactReduxContext` for consuming the redux store, which means it currently does not support custom context for Redux. Support for this might be added in the future.

* chore(*): update various dependencies
* chore(ConnectedUIRouter): better types
* feat(ConnectedUIRouter): support react-redux@^6.x
* chore(*): update deps and tests

New component implementation requires at least react@16 and the new enzyme adapter for it.
Updated the tests to use the new Redux Context api, as well as rely on public uirouter APIs instead of component internals to be more reliable.

* chore(travis): update node version
* fix: update typings to fix typescript errors
* fix: update all dependencies
* fix: update webpack config for latest webpack

Co-authored-by: Chris Thielen <christopherthielen@users.noreply.github.com>